Refactor ppocr parser into ClassificationSequenceParser to be in line with new parser definitions.
